refactor(finance): extract ReviewStep hooks and sub-components [tb-368]
Extract from ReviewStep.tsx (1184→386 LOC):
- useTransactionReview: local transaction state, view mode, scroll tracking
- useProposalGeneration: AI analysis, proposal dialog state
- useTransactionEditing: edit/save/cancel handlers
- useBulkAssignment: accept-all, create-and-assign, entity creation

Extract 4 tab sub-components to review/:
  MatchedTab, UncertainTab, FailedTab, SkippedTab
